Replace sink and GetError with segmentor callbacks
